Split astro_rewrite into smaller files and subpkg
This improves the maintainability of the code by making things
actually readable. Furthermore, it replaces the cruft that
was used to calculate moon phases with a much simpler method.
The solar returns and lunar returns will have similar updates.

As a result, the listbox for moon phases has a lot less entries
since knowing the specific time of when a phase starts is more
important to the budding spellcaster than simply knowing what phase
is on what day. The calendar will be updated to reflect this.

A unit test was also added to make sure that any edge cases
that cause the moon phase approximation to break can be
repeatedly and reliably tested to restore such functionality.

There's also a version bump.