Add method to observer for averages
This is useful for combine charts.

Also, note:
Since the CLI tool in development (not in repo) focuses on
churning out batches of astrological chart data, I'll likely
have to split the core library from the GUI to make it clear
that the core library is simply a buncha stuff that helps with
all kinds of astrology.