Simplify initalizing the zodiac
Before this commit, ZODIAC was a giant dictionary filled manually with
the appropriate values. All of those values fit certain patterns,
though.

So to simplify readability and maintenance,
ZODIAC was renamed to Zodiac and all of the Zodiac entries are Enums.
This goes for some of their properties too - like the elements and their
modes.