Calc sunset and next sunrise relative to sunrise
After reading the swisseph documentation more carefully,
I noticed that the day parameter is the time to start searching after.
To ensure times are always in order, we pass the julian day
value of the correct estimated sunrise to the other calculations.

This fixes the Antartica/McMurdo bug discovered by the unit test.